refactor: move recaps to synthesis/recaps.rs
