+++ crates/origin-core/src/eval/lifecycle.rs
@@ -786,7 +786,9 @@ async fn run_lifecycle_phases(
     // --- Phase 4: Insights (LLM required) ---
     let phase_start = std::time::Instant::now();
     if let Some(llm_ref) = llm {
-        let _ = crate::refinery::generate_recaps(db, Some(llm_ref), &prompts, &refinery_cfg).await;
+        let _ =
+            crate::synthesis::recaps::generate_recaps(db, Some(llm_ref), &prompts, &refinery_cfg)
+                .await;
         let _ = crate::refinery::generate_decision_logs(db, Some(llm_ref), &prompts, &refinery_cfg)
             .await;
     }
