+++ crates/origin-core/src/synthesis/mod.rs
@@ -1,0 +1,8 @@
+// SPDX-License-Identifier: Apache-2.0
+//! Synthesis phases — derived views over raw memories.
+//!
+//! These phases generate compact summaries from clusters of memories: recaps
+//! (chronological digests) and decision_logs (decision tracebacks). They run
+//! periodically rather than on each ingest.
+
+pub mod recaps;
